std:: etc

+++ src/ofxAsio/Message.h
@@ -74,8 +74,8 @@ struct Pixels_ {
 		uint16_t width;
 		uint16_t height;
 		uint16_t pixelFormat;
-		
-		uint8_t userData[2];
+
+		uint16_t userData[2];
 
 	};
 };
